[caffe2] Fix alias analysis for quantization compression ops
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74169

Alias DB was being way too conservative about the semantics of exported Caffe2 ops - it thought some pure functions were writing to their inputs, which caused `ReplaceWithMaybeCopy` to fail. This in turn lead to a huge decrease in out variant coverage and regressions in many models.

I've extended the export macro to let the user specify an `AliasAnalysisKind` and marked all of the quantization compression ops as pure functions.
ghstack-source-id: 151283346

Differential Revision: [D34733630](https://our.internmc.facebook.com/intern/diff/D34733630/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D34733630/)!